fix(emails): move emails folder into components, fix styling
