+++ sim/components/emails/base-styles.ts
