+++ sim/components/emails/email-icons.tsx
