+++ sim/package.json
@@ -17,7 +17,7 @@
     "test": "vitest run",
     "test:watch": "vitest",
     "test:coverage": "vitest run --coverage",
-    "email:dev": "email dev --dir app/emails",
+    "email:dev": "email dev --dir components/emails",
     "cli:build": "npm run build -w packages/simstudio",
     "cli:dev": "npm run build -w packages/simstudio && cd packages/simstudio && node ./dist/index.js",
     "cli:publish": "cd packages/simstudio && npm publish",
